@@ -4,51 +4,53 @@
     "minimum-stability": "dev",
     "prefer-stable": true,
     "require": {
-        "php": ">=8.0.2",
+        "php": ">=8.1",
         "ext-ctype": "*",
         "ext-iconv": "*",
         "composer/package-versions-deprecated": "1.11.99.1",
-        "doctrine/annotations": "^1.11",
-        "doctrine/doctrine-bundle": "^2.1",
-        "doctrine/doctrine-migrations-bundle": "^3.0",
-        "doctrine/orm": "^2.8",
+        "cweagans/composer-patches": "^1.7",
+        "doctrine/doctrine-bundle": "^2.7",
+        "doctrine/doctrine-migrations-bundle": "^3.2",
+        "doctrine/orm": "^2.14",
         "easycorp/easyadmin-bundle": "^4.0",
+        "gedmo/doctrine-extensions": "^3.10",
         "pagerfanta/doctrine-orm-adapter": "^3.3",
         "pagerfanta/twig": "^3.3",
         "sensio/framework-extra-bundle": "^6.2",
-        "symfony/asset": "6.0.*",
-        "symfony/cache": "6.0.*",
-        "symfony/config": "6.0.*",
-        "symfony/console": "6.0.*",
-        "symfony/debug-bundle": "6.0.*",
-        "symfony/dotenv": "6.0.*",
+        "symfony/asset": "6.1.*",
+        "symfony/cache": "6.1.*",
+        "symfony/config": "6.1.*",
+        "symfony/console": "6.1.*",
+        "symfony/debug-bundle": "6.1.*",
+        "symfony/dependency-injection": "6.1.*",
+        "symfony/dotenv": "6.1.*",
         "symfony/flex": "^2.0.0",
-        "symfony/form": "6.0.*",
-        "symfony/framework-bundle": "6.0.*",
-        "symfony/http-kernel": "6.0.*",
-        "symfony/mime": "6.0.*",
+        "symfony/form": "6.1.*",
+        "symfony/framework-bundle": "6.1.*",
+        "symfony/http-kernel": "6.1.*",
+        "symfony/mime": "6.1.*",
         "symfony/monolog-bundle": "^3.0",
-        "symfony/security-bundle": "6.0.*",
+        "symfony/security-bundle": "6.1.*",
         "symfony/security-csrf": "5.4.*",
-        "symfony/serializer": "6.0.*",
-        "symfony/stopwatch": "6.0.*",
-        "symfony/translation": "6.0.*",
-        "symfony/twig-bundle": "6.0.*",
+        "symfony/serializer": "6.1.*",
+        "symfony/stopwatch": "6.1.*",
+        "symfony/translation": "6.1.*",
+        "symfony/twig-bundle": "6.1.*",
         "symfony/ux-chartjs": "^2.5",
         "symfony/ux-turbo": "^2.5",
-        "symfony/validator": "6.0.*",
-        "symfony/var-dumper": "6.0.*",
+        "symfony/validator": "6.1.*",
+        "symfony/var-dumper": "6.1.*",
         "symfony/web-profiler-bundle": "^5.4",
         "symfony/webpack-encore-bundle": "^1.11",
-        "symfony/yaml": "6.0.*",
+        "symfony/yaml": "6.1.*",
         "twig/extra-bundle": "^2.12|^3.0",
         "twig/twig": "^2.12|^3.0"
     },
     "require-dev": {
-        "symfony/maker-bundle": "^1.25",
-        "symfony/phpunit-bridge": "^6.0",
-        "symfony/var-dumper": "6.0.*",
-        "symfony/debug-bundle": "6.0.*"
+        "rector/rector": "^0.15.2",
+        "symfony/debug-bundle": "6.1.*",
+        "symfony/maker-bundle": "^1.48",
+        "symfony/var-dumper": "6.1.*"
     },
     "config": {
         "optimize-autoloader": true,
@@ -59,7 +61,8 @@
         "allow-plugins": {
             "composer/package-versions-deprecated": true,
             "symfony/flex": true,
-            "symfony/runtime": true
+            "symfony/runtime": true,
+            "cweagans/composer-patches": true
         }
     },
     "autoload": {
@@ -99,7 +102,7 @@
     "extra": {
         "symfony": {
             "allow-contrib": false,
-            "require": "6.0.*"
+            "require": "6.1.*"
         }
     }
 }
