@@ -3,92 +3,80 @@
 namespace App\Controller;
 
 use App\Entity\Area;
-use App\Form\AreaType;
+use App\Form\Area1Type;
 use App\Repository\AreaRepository;
-use App\Repository\RockRepository;
-use App\Service\MarkdownHelper;
-use Doctrine\ORM\EntityManagerInterface;
-use Doctrine\Persistence\ManagerRegistry;
-use Psr\Log\LoggerInterface;
 use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\Response;
 use Symfony\Component\Routing\Annotation\Route;
+use Doctrine\ORM\EntityManagerInterface;
 
-/**
- * @Route("/area")
- */
+
+#[Route('/area')]
 class AreaController extends AbstractController
 {
-    /**
-     * @Route("/", name="area_index", methods={"GET"})
-     */
+    #[Route('/', name: 'app_area_index', methods: ['GET'])]
     public function index(AreaRepository $areaRepository): Response
     {
         return $this->render('area/index.html.twig', [
-            'areas' => $areaRepository->findAllAreasAlphabetical(),
+            'areas' => $areaRepository->findAll(),
         ]);
     }
 
-    /**
-     * @Route("/new", name="area_new", methods={"GET","POST"})
-     */
-    public function new(ManagerRegistry $doctrine, Request $request, AreaRepository $areaRepository): Response
+    #[Route('/new', name: 'app_area_new', methods: ['GET', 'POST'])]
+    public function new(Request $request, EntityManagerInterface $entityManager): Response
     {
         $area = new Area();
-        $form = $this->createForm(AreaType::class, $area);
+        $form = $this->createForm(Area1Type::class, $area);
         $form->handleRequest($request);
 
         if ($form->isSubmitted() && $form->isValid()) {
-            $entityManager = $doctrine->getManager();
             $entityManager->persist($area);
             $entityManager->flush();
 
-            return $this->redirectToRoute('area_index');
+            return $this->redirectToRoute('app_area_index', [], Response::HTTP_SEE_OTHER);
         }
 
-        return $this->render('area/new.html.twig', [
+        return $this->renderForm('area/new.html.twig', [
             'area' => $area,
-            'form' => $form->createView(),
-            'areas' => $areaRepository->findAllAreasAlphabetical(),
+            'form' => $form,
         ]);
     }
 
-    /**
-     * @Route("/{id}/edit", name="area_edit", methods={"GET","POST"})
-     */
-    public function edit(ManagerRegistry $doctrine, Request $request, Area $area, AreaRepository $areaRepository, RockRepository $rockRepository): Response
+    #[Route('/{id}', name: 'app_area_show', methods: ['GET'])]
+    public function show(Area $area): Response
     {
-        
-        $form = $this->createForm(AreaType::class, $area);
+        return $this->render('area/show.html.twig', [
+            'area' => $area,
+        ]);
+    }
+
+    #[Route('/{id}/edit', name: 'app_area_edit', methods: ['GET', 'POST'])]
+    public function edit(Request $request, Area $area, EntityManagerInterface $entityManager): Response
+    {
+        $form = $this->createForm(Area1Type::class, $area);
         $form->handleRequest($request);
 
         if ($form->isSubmitted() && $form->isValid()) {
-            $doctrine->getManager()->flush();
+            $entityManager->flush();
 
-            $this->addFlash('success', 'Gebiet wurde erfolgreich aktualisiert');
-
-            return $this->redirectToRoute('area_index');
+            return $this->redirectToRoute('app_area_index', [], Response::HTTP_SEE_OTHER);
         }
 
-        return $this->render('area/edit.html.twig', [
+        return $this->renderForm('area/edit.html.twig', [
             'area' => $area,
-            'form' => $form->createView(),
-            'areas' => $areaRepository->findAllAreasAlphabetical(),
+            'form' => $form,
         ]);
     }
 
-    /**
-     * @Route("/{id}", name="area_delete", methods={"DELETE"})
-     */
-    public function delete(Request $request, Area $area): Response
+    #[Route('/{id}', name: 'app_area_delete', methods: ['POST'])]
+    public function delete(Request $request, Area $area, EntityManagerInterface $entityManager): Response
     {
         if ($this->isCsrfTokenValid('delete'.$area->getId(), $request->request->get('_token'))) {
-            $entityManager = $doctrine->getManager();
             $entityManager->remove($area);
             $entityManager->flush();
         }
 
-        return $this->redirectToRoute('area_index');
+        return $this->redirectToRoute('app_area_index', [], Response::HTTP_SEE_OTHER);
     }
 }
