@@ -6,7 +6,7 @@ use App\Entity\Rock;
 use App\Entity\Routes;
 use App\Entity\Topo;
 use App\Entity\User;
-use App\Entity\Galerie;
+use App\Entity\Photos;
 use App\Entity\Videos;
 use App\Repository\AreaRepository;
 use App\Repository\RockRepository;
@@ -95,9 +95,9 @@ class DashboardController extends AbstractDashboardController
         yield MenuItem::linkToCrud('Felsen', 'fa fa-home', Rock::class);
         yield MenuItem::linkToCrud('Touren', 'fa fa-home', Routes::class);
         yield MenuItem::linkToCrud('Topos', 'fa fa-home', Topo::class);
-        yield MenuItem::linkToCrud('Photos', 'fa fa-camera-retro', Galerie::class);
+        yield MenuItem::linkToCrud('Photos', 'fa fa-camera-retro', Photos::class);
         yield MenuItem::linkToCrud('Videos', 'fa fa-video', Videos::class);
-        yield MenuItem::linkToCrud('User', 'fa fa-user', User::class);
+        yield MenuItem::linkToCrud('User', 'fa fa-user', User::class)->setPermission('ROLE_SUPER_ADMIN');
         yield MenuItem::section('Live Seite');
         yield MenuItem::linkToUrl('munichclimbs', 'fa fa-link', 'https://munichclimbs.de')->setLinkTarget('_blank');
         yield MenuItem::section('Tools');
