@@ -206,11 +206,7 @@ class RockType extends AbstractType
     public function configureOptions(OptionsResolver $resolver)
     {
         $resolver->setDefaults([
-            'data_class' => Rock::class,
-            'attr' => [
-                'class' => 'js-user-autocomplete',
-                'data-autocomplete-url' => $this->router->generate('admin_utility_users')
-            ]
+            'data_class' => Rock::class
         ]);
     }
 }
